pyproject.toml
[tool.poetry]
name = "kg3_info_bot"
version = "0.1.0"
description = ""
authors = ["Oleksand Varchenko <oleksandr.varchenko@globallogic.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
debugpy = "^1.5.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
